Fix: potential overflow in ssr (#1600)

+++ transport/ssr/protocol/auth_chain_a.go
@@ -278,7 +278,7 @@ func getRandStartPos(length int, random *tools.XorShift128Plus) int {
 	if length == 0 {
 		return 0
 	}
-	return int(random.Next()%8589934609) % length
+	return int(int64(random.Next()%8589934609) % int64(length))
 }
 
 func (a *authChainA) getRandLength(length int, lastHash []byte, random *tools.XorShift128Plus) int {
